izlabots
novērsts int overlow

+++ 5majasdarbs.c
@@ -4,12 +4,12 @@
 //Vienkārša programma, kas pieprasa keyboard input (stdin) un izdara reizināšanas darbību un izvada atbildi
 
 int main(void){
-    int a,b;
+    long long a,b;
 
     printf("Reizinasanas operacija, kur starp skaitliem vajag atstarpi: \n pirmais un otrais skaitlis:");fflush(stdout);
-    scanf("%d %d", &a, &b);
+    scanf("%lld %lld", &a, &b);
 
-    int reizinajums = a * b;
+    long long reizinajums = a * b;
     
-    printf("\n Atbilde: %d", reizinajums);
+    printf("\n Atbilde: %lld", reizinajums);
 }
